feat: Allow duplicate server names  (#1082)
* feat: Allow duplicate server names & update to GormV2 for manual migrations

* feat: Remove the windows_loader

* tidying up

* fix: An issue with how the transaction was being applied

* fix: Made the migration infallible

* fix: put the import back :|

* chore: Revert a change that snuck in

* fix: remove a nested transaction

* fix: add some code for scopes back in

* fix: Corrected, again, ports to uint16

* fix: Removed the transaction & noted the caveat behind the changes (gorm changed things abd broke the logic

* fix: Remove the user from the users table

* style: Move 'gorm.ErrRecordNotFound' to == from errors.is

* refactor(?): Move pages back to uints

* docs: Correct the current type for ID's

* fix: Rollback/Commit inside the transaction middleware

+++ cmd/run.go
@@ -168,14 +168,14 @@ func panel(ch chan error) {
 			if len(hostParts) == 2 {
 				port, err := strconv.Atoi(hostParts[1])
 				if err == nil {
-					create.PublicPort = uint(port)
-					create.PrivatePort = uint(port)
+					create.PublicPort = uint16(port)
+					create.PrivatePort = uint16(port)
 				}
 			}
 			if len(sftpParts) == 2 {
 				port, err := strconv.Atoi(sftpParts[1])
 				if err == nil {
-					create.SFTPPort = uint(port)
+					create.SFTPPort = uint16(port)
 				}
 			}
